refactor: move DASH manifest stuff out of `VideoInfo`
This makes it simpler to use these functions elsewhere.
